# backend/docs/swagger.yaml
definitions:
  dto.UserDTO:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      email:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        default: ACTIVE
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: integer
      userId:
        type: integer
      username:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      email:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        type: string
      teacherId:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: integer
      userId:
        type: integer
      username:
        type: string
    type: object
info:
  contact:
    email: pham.huy.19951126@gmail.com
    name: Huy Pham
  description: This is E-learning Api
  termsOfService: none
  title: E-learning Api
  version: "1.0"
paths:
  /admin/user/create:
    post:
      consumes:
      - application/json
      description: Create a new user with role TEACHER or STUDENT. Only accessible
        by ADMIN.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User info to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input or role
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user (teacher or student)
      tags:
      - Admin Users
  /users:
    get:
      consumes:
      - application/json
      description: get string by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      summary: Show an account
      tags:
      - Users
swagger: "2.0"